+++ SBBMobileDesignSystemSwiftUI/Colors/SBBColors.swift
@@ -6,6 +6,13 @@ import Foundation
 import SwiftUI
 
 public extension Color {
+    
+    /// The primary color set in SBBTheme which is used for many UI elements (e.g. NavigationBar background, SBBPrimaryButton).
+    static var sbbPrimary: Color = Color.sbbColor(.red)
+    
+    /// The secondary color set in SBBTheme which is used for pressed states/backgrounds of some UI elements (e.g. SBBPrimaryButton in pressed state).
+    static var sbbSecondary: Color = Color.sbbColor(.red150)
+    
     /**
      Returns a SBB Color.
      
@@ -13,6 +20,11 @@ public extension Color {
         - name: The SBBColorName of the color to be returned.
      */
     static func sbbColor(_ name: SBBColorName) -> Color {
+        if name == .primary {
+            return sbbPrimary
+        } else if name == .secondary {
+            return sbbSecondary
+        }
         return Color(UIColor.sbbColor(name))
     }
 }