+++ SBBMobileDesignSystemSwiftUIDemo/Views/DetailViews/MapMarkerDemo.swift
@@ -11,7 +11,7 @@ struct MapMarkerDemo: View {
     public var body: some View {
         ScrollView(showsIndicators: false) {
             VStack(spacing: 24) {
-                SBBInfoView(image: Image(sbbName: "circle-information", size: .small), text: Text("There are 3 available map marker styles: .red, .blue and .black. Use .blue style for pictograms."))
+                SBBInfoView(image: Image(sbbName: "circle-information", size: .small), text: Text("There are 3 available map marker styles: .primary, .blue and .black. Use .blue style for pictograms."))
                 HStack(spacing: 16) {
                     SBBMapMarker(icon: Image(sbbName: "construction"))
                     SBBMapMarker(icon: Image(sbbName: "Zug_r"), style: .blue)